Move tokenizer into new file.
