+++ resgenclass.cpp
@@ -41,6 +41,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include <unistd.h>
 #endif
 
+#include "enttokenizer.h"
 #include "hltypes.h"
 #include "resgenclass.h"
 #include "resgen.h"
