+++ util.h
@@ -20,7 +20,6 @@ const char PATH_SEPARATOR = '/';
 #endif
 
 #include <algorithm>
-#include <cstddef>
 #include <exception>
 #include <string>
 #include <vector>
@@ -103,190 +102,6 @@ public:
 
 private:
 	int charNum;
-};
-
-template <char Delimiter>
-class Tokenizer
-{
-public:
-	Tokenizer(std::string& str_)
-		: str(str_)
-		, currentPtr(&str[0])
-		, strEnd(currentPtr + str.length())
-	{
-	}
-
-	INLINE bool Next()
-	{
-		if(!currentPtr)
-		{
-			return false;
-		}
-
-		while(currentPtr != strEnd)
-		{
-			if(*currentPtr == Delimiter)
-			{
-				*currentPtr = 0;
-				currentPtr++;
-				return true;
-			}
-
-			currentPtr++;
-		}
-
-		throw ParseException("Found end of data while parsing string");
-	}
-
-protected:
-	Tokenizer(const Tokenizer &other);
-	Tokenizer& operator=(const Tokenizer &other);
-
-	std::string str;
-
-	// Current position in string
-	char* currentPtr;
-
-	// Pointer to char after last
-	char* strEnd;
-};
-
-class EntTokenizer : public Tokenizer<'\"'>
-{
-public:
-	typedef std::pair<const char*, const char*> KeyValuePair;
-
-	EntTokenizer(std::string& str_)
-		: Tokenizer(str_)
-		, bInBlock(false)
-		, blocksRead(0)
-	{
-	}
-
-	int GetCharNum() const
-	{
-		if(!currentPtr)
-		{
-			return -1;
-		}
-
-		return currentPtr - &str[0];
-	}
-
-	int GetNumBlocksRead() const
-	{
-		return blocksRead;
-	}
-
-	ptrdiff_t GetLatestValueLength() const
-	{
-		// currentPtr points to char after NUL
-		const char* const end = (currentPtr ? currentPtr - 1 : strEnd);
-		return end - pair.second;
-	}
-
-	const KeyValuePair* NextPair()
-	{
-		// Have we finished parsing?
-		if(!currentPtr)
-		{
-			return NULL;
-		}
-
-		// Skip end of line
-		if(!NextKV())
-		{
-			// If we've run out key-values, make sure we are in an accept state
-			if(bInBlock)
-			{
-				throw ParseException("Reached end of entity data while inside a block.", GetCharNum());
-			}
-
-			return NULL;
-		}
-
-		// Read key
-		pair.first = currentPtr;
-
-		if(!Next())
-		{
-			throw ParseException("Failed to parse key of key-value pair.", GetCharNum());
-		}
-
-		// Ignore space between key/value
-		ParseKVSeparator();
-
-		pair.second = currentPtr;
-
-		if(!Next())
-		{
-			throw ParseException("Failed to parse value of key-value pair.", GetCharNum());
-		}
-
-		return &pair;
-	}
-
-private:
-	void ParseKVSeparator()
-	{
-		while(currentPtr != strEnd)
-		{
-			if(*currentPtr == '\"')
-			{
-				*currentPtr = 0;
-				currentPtr++;
-				return;
-			}
-			else if(*currentPtr != ' ')
-			{
-				throw ParseException("Found non-whitespace between key and value.", GetCharNum());
-			}
-
-			currentPtr++;
-		}
-
-		throw ParseException("Failed to parse key-value pair", GetCharNum());
-	}
-
-	bool NextKV()
-	{
-		while(currentPtr != strEnd)
-		{
-			switch(*currentPtr)
-			{
-				case '\"':
-					*currentPtr = 0;
-					currentPtr++;
-					return true;
-				case '{':
-					if(bInBlock)
-					{
-						throw ParseException("Unexpected start of block.", GetCharNum());
-					}
-					bInBlock = true;
-					break;
-				case '}':
-					if(!bInBlock)
-					{
-						throw ParseException("Unexpected end of block.", GetCharNum());
-					}
-					bInBlock = false;
-					blocksRead++;
-					break;
-				default:
-					break;
-			}
-
-			currentPtr++;
-		}
-
-		currentPtr = NULL;
-		return false;
-	}
-
-	KeyValuePair pair;
-	bool bInBlock;
-	int blocksRead;
 };
